Complete code
The code to reproduce our upcoming article about income inequalities in life expectancy among older people in Sweden is available here. All 79 Stata do-files must be downloaded in a single folder on your computer, together with 3 empty subfolders: 
1. "Final estimates"
2. "Life expectancy estimates (appended tables)"
3. "Life expectancy estimates (individual tables)"

Once this is done, simply run the file "Pipeline.do" in Stata. It will create the necessary variables, split the main dataset in year-specific data tables, and execute all the do-files automatically.

The file "dataset.dta" is a mock version of the actual dataset, which cannot be made publicly available for privacy reasons.